add isolated modules option, fix all errors
