+++ .github/workflows/main.yml
@@ -18,7 +18,7 @@ jobs:
               with:
                   node-version: ${{ matrix.node-version }}
             - name: install modules
-              run: npm install
+              run: npm install --force
             - name: build production project
               run: npm run build:prod
               if: always()
