+++ json-server/db.json
@@ -1448,6 +1448,12 @@
       "userId": "1",
       "text": "OMG!",
       "id": "LnsYqEg"
+    },
+    {
+      "articleId": "10",
+      "userId": "1",
+      "text": "fpjewfew",
+      "id": "-JWSHWF"
     }
   ],
   "users": [
