+++ package.json
@@ -84,7 +84,7 @@
         "reg-cli": "^0.17.6",
         "sass": "^1.49.9",
         "sass-loader": "^12.6.0",
-        "storybook-addon-mock": "^4.1.0",
+        "storybook-addon-mock": "^2.4.1",
         "style-loader": "^3.3.1",
         "stylelint": "^14.5.3",
         "stylelint-config-standard-scss": "^3.0.0",
