+++ src/app/providers/StoreProvider/index.ts
@@ -1,16 +1,8 @@
-import {
+export type {
     ReduxStoreWithManager,
     StateSchema,
     ThunkConfig
 } from './config/StateSchema';
-import { AppDispatch, createReduxStore } from './config/store';
-import { StoreProvider } from './ui/StoreProvider';
-
-export {
-    StoreProvider,
-    createReduxStore,
-    StateSchema,
-    ReduxStoreWithManager,
-    AppDispatch,
-    ThunkConfig
-};
+export type { AppDispatch } from './config/store';
+export { createReduxStore } from './config/store';
+export { StoreProvider } from './ui/StoreProvider';
