+++ src/entities/Article/index.ts
@@ -1,8 +1,12 @@
 export { ArticleDetails } from './ui/ArticleDetails/ArticleDetails';
-export { Article, ArticleView } from './model/types/article';
+export type { Article } from './model/types/article';
 export type { ArticleDetailsSchema } from './model/types/articleDetailsSchema';
 export { ArticleList } from './ui/ArticleList/ArticleList';
-export { ArticleSortField, ArticleType } from './model/types/article';
+export {
+    ArticleSortField,
+    ArticleType,
+    ArticleView
+} from './model/consts/articleConsts';
 export { ArticleSortSelector } from './ui/ArticleSortSelector/ArticleSortSelector';
 export { ArticleTypeTabs } from './ui/ArticleTypeTabs/ArticleTypeTabs';
 export { getArticleDetailsData } from './model/selectors/articleDetails';
