+++ src/entities/Article/model/consts/articleConsts.ts
@@ -1,0 +1,23 @@
+export enum ArticleSortField {
+    VIEWS = 'views',
+    TITLE = 'title',
+    CREATED = 'createdAt'
+}
+
+export enum ArticleBlockType {
+    CODE = 'CODE',
+    TEXT = 'TEXT',
+    IMAGE = 'IMAGE'
+}
+
+export enum ArticleType {
+    ALL = 'ALL',
+    IT = 'IT',
+    SCIENCE = 'SCIENCE',
+    ECONOMICS = 'ECONOMICS'
+}
+
+export enum ArticleView {
+    BIG = 'BIG',
+    SMALL = 'SMALL'
+}
