+++ src/entities/Article/ui/ArticleDetails/ArticleDetails.stories.tsx
@@ -2,11 +2,11 @@ import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { StoreDecorator } from 'shared/config/storybook/StoreDecorator/StoreDecorator';
 import { ThemeDecorator } from 'shared/config/storybook/ThemeDecorator/ThemeDecorator';
 import { Theme } from 'app/providers/ThemeProvider';
+import { Article } from '../../model/types/article';
 import {
     ArticleBlockType,
-    ArticleType,
-    Article
-} from '../../model/types/article';
+    ArticleType
+} from '../../model/consts/articleConsts';
 
 import { ArticleDetails } from './ArticleDetails';
