+++ src/entities/Article/ui/ArticleListItem/ArticleListItem.tsx
@@ -10,12 +10,11 @@ import { Avatar } from 'shared/ui/Avatar/Avatar';
 import { RoutePath } from 'shared/config/routeConfig/routeConfig';
 import { AppLink } from 'shared/ui/AppLink/AppLink';
 import cls from './ArticleListItem.module.scss';
+import { Article, ArticleTextBlock } from '../../model/types/article';
 import {
-    Article,
     ArticleBlockType,
-    ArticleTextBlock,
     ArticleView
-} from '../../model/types/article';
+} from '../../model/consts/articleConsts';
 import { ArticleTextBlockComponent } from '../ArticleTextBlockComponent/ArticleTextBlockComponent';
 
 interface ArticleListItemProps {
