+++ src/entities/Article/ui/ArticleListItem/ArticleListItemSkeleton.tsx
@@ -3,7 +3,7 @@ import { classNames } from 'shared/lib/helpers/classNames';
 import { Card } from 'shared/ui/Card/Card';
 import { Skeleton } from 'shared/ui/Skeleton/Skeleton';
 import cls from './ArticleListItem.module.scss';
-import { ArticleView } from '../../model/types/article';
+import { ArticleView } from '../../model/consts/articleConsts';
 
 interface ArticleListItemSkeletonProps {
     className?: string;
