+++ src/entities/Comment/index.ts
@@ -1,2 +1,2 @@
 export { CommentList } from './ui/CommentList/CommentList';
-export { Comment } from './model/types/comment';
+export type { Comment } from './model/types/comment';
