+++ src/entities/Counter/index.ts
@@ -1,5 +1,3 @@
-import { counterSlice } from './model/slice/counterSlice';
-import { CounterSchema } from './model/types/counterSchema';
-import { Counter } from './ui/Counter';
-
-export { Counter, CounterSchema, counterSlice };
+export { counterSlice } from './model/slice/counterSlice';
+export type { CounterSchema } from './model/types/counterSchema';
+export { Counter } from './ui/Counter';
