+++ src/entities/Profile/index.ts
@@ -1,2 +1,2 @@
-export { Profile } from './model/types/profile';
+export type { Profile } from './model/types/profile';
 export { ProfileCard } from './ui/ProfileCard/ProfileCard';
