+++ src/entities/User/model/consts/userConsts.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+    ADMIN = 'ADMIN',
+    MANAGER = 'MANAGER',
+    USER = 'USER'
+}
