+++ src/entities/User/model/selectors/roleSetectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { StateSchema } from 'app/providers/StoreProvider';
-import { UserRole } from '../types/user';
+import { UserRole } from '../consts/userConsts';
 
 export const getUserRoles = (state: StateSchema) =>
     state?.user?.authData?.roles;
