+++ src/entities/User/model/types/user.ts
@@ -1,8 +1,4 @@
-export enum UserRole {
-    ADMIN = 'ADMIN',
-    MANAGER = 'MANAGER',
-    USER = 'USER'
-}
+import { UserRole } from '../consts/userConsts';
 
 export interface User {
     id: string;
