+++ src/features/AddCommentForm/index.ts
@@ -1,2 +1,2 @@
-export { AddCommentFormSchema } from './model/types/AddCommentForm';
+export type { AddCommentFormSchema } from './model/types/AddCommentForm';
 export { AddCommentFormAsync as AddCommentForm } from './ui/AddCommentForm/AddCommentForm.async';
