+++ src/features/AuthByUserName/index.ts
@@ -1,4 +1,2 @@
-import { LoginSchema } from './model/types/LoginSchema';
-import { LoginModal } from './ui/LoginModal/LoginModal';
-
-export { LoginModal, LoginSchema };
+export type { LoginSchema } from './model/types/LoginSchema';
+export { LoginModal } from './ui/LoginModal/LoginModal';
