+++ src/features/AuthByUserName/model/services/loginByUsername/loginByUsername.ts
@@ -29,7 +29,6 @@ export const loginByUsername = createAsyncThunk<
 
         return response.data;
     } catch (e) {
-        console.log(e);
         return rejectWithValue('error');
     }
 });
