+++ src/features/UI/index.ts
@@ -1,4 +1,4 @@
-export { UISchema } from './model/types/UISchema';
+export type { UISchema } from './model/types/UISchema';
 
 export { getUIScrollByPath } from './model/selectors/ui';
 export { uiActions, uiReducer } from './model/slices/UISlice';
