+++ src/features/editableProfileCard/model/consts/consts.ts
@@ -1,0 +1,11 @@
+export enum ValidateProfileError {
+    INCORRECT_FIRSTNAME = 'INCORRECT_FIRSTNAME',
+    INCORRECT_LASTNAME = 'INCORRECT_LASTNAME',
+    INCORRECT_AGE = 'INCORRECT_AGE',
+    INCORRECT_USERNAME = 'INCORRECT_USERNAME',
+    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
+    INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
+    INCORRECT_CITY = 'INCORRECT_CITY',
+    NO_DATA = 'NO_DATA',
+    SERVER_ERROR = 'SERVER_ERROR'
+}
