+++ src/pages/ArticleDetailsPage/index.ts
@@ -1,4 +1,4 @@
 export { ArticleDetailsPageAsync as ArticleDetailsPage } from './ui/ArticleDetailsPage/ArticleDetailsPage.async';
-export { ArticleDetailsCommentSchema } from './model/types/ArticleDetailsCommentSchema';
-export { ArticleDetailsPageRecommendationsSchema } from './model/types/articleDetailsPageRecommendationsSchema';
-export { ArticleDetailsPageSchema } from './model/types/index';
+export type { ArticleDetailsCommentSchema } from './model/types/ArticleDetailsCommentSchema';
+export type { ArticleDetailsPageRecommendationsSchema } from './model/types/articleDetailsPageRecommendationsSchema';
+export type { ArticleDetailsPageSchema } from './model/types/index';
