+++ src/pages/ArticlesPage/index.ts
@@ -1,2 +1,2 @@
 export { ArticlesPageAsync as ArticlesPage } from './ui/ArticlesPage/ArticlesPage.async';
-export { ArticlePageSchema } from './model/types/articlesPageSchema';
+export type { ArticlePageSchema } from './model/types/articlesPageSchema';
