+++ src/pages/ArticlesPage/model/services/fetchArticlesList/fetchArticlesList.test.ts
@@ -3,7 +3,7 @@ import { Article } from 'entities/Article';
 import {
     ArticleBlockType,
     ArticleType
-} from 'entities/Article/model/types/article';
+} from 'entities/Article/model/consts/articleConsts';
 import { TestAsyncThunk } from 'shared/lib/tests/TestAsyncThunk/TestAsyncThunk';
 import { fetchArticlesList } from './fetchArticlesList';
