+++ src/pages/ArticlesPage/model/slices/articlesPageSlice.test.ts
@@ -2,7 +2,7 @@ import { Article, ArticleView } from 'entities/Article';
 import {
     ArticleBlockType,
     ArticleType
-} from 'entities/Article/model/types/article';
+} from 'entities/Article/model/consts/articleConsts';
 import { fetchArticlesList } from '../services/fetchArticlesList/fetchArticlesList';
 import { ArticlePageSchema } from '../types/articlesPageSchema';
 import { articlesPageActions, articlesPageReducer } from './articlesPageSlice';
