+++ src/widgets/Sidebar/index.ts
@@ -1,3 +1,1 @@
-import { Sidebar } from './ui/Sidebar/Sidebar';
-
-export { Sidebar };
+export { Sidebar } from './ui/Sidebar/Sidebar';
